diff --git a/package-lock.json b/package-lock.json
removed axios as dependency; updated to 0.21.1 to avoid vulnerability

diff --git a/starttlist.sh b/starttlist.sh

added update wrapper

+++ starttlist.sh
@@ -1,5 +1,8 @@
 #!/bin/sh
-cat dot-env-todo >.env
-cp -v todo-data.json data.json
+
+if [ -e UPDATE ]; then
+	cat dot-env-todo >.env
+	cp -v todo-data.json data.json
+fi
 npm start
